languages API fixes. Performance fixes still needed
