+++ crates/languages/src/rust/mod.rs
@@ -10,7 +10,7 @@ pub(crate) mod queries;
 use crate::extraction_framework::{GenericExtractor, LanguageConfigurationBuilder};
 use crate::transport::EntityData;
 use crate::Extractor;
-use codesearch_core::entities::{CodeEntityBuilder, Language};
+use codesearch_core::entities::Language;
 use codesearch_core::error::Result;
 use codesearch_core::CodeEntity;
 use std::path::Path;
@@ -56,48 +56,6 @@ impl RustExtractor {
 
         GenericExtractor::new(config_ptr)
     }
-
-    /// Convert EntityData to CodeEntity
-    fn convert_to_code_entity(&self, entity: EntityData, file_path: &Path) -> CodeEntity {
-        let entity_type = entity.variant.entity_type();
-        let metadata = entity.variant.into_metadata();
-        let signature = entity.variant.extract_signature();
-        let location = entity.location.clone();
-
-        CodeEntityBuilder::default()
-            .entity_id(format!("{}#{}", file_path.display(), entity.qualified_name))
-            .name(entity.name)
-            .qualified_name(entity.qualified_name)
-            .entity_type(entity_type)
-            .location(entity.location)
-            .visibility(entity.visibility)
-            .documentation_summary(entity.documentation)
-            .content(entity.content)
-            .dependencies(entity.dependencies)
-            .metadata(metadata)
-            .signature(signature)
-            .language(Language::Rust)
-            .build()
-            .unwrap_or_else(|e| {
-                tracing::error!("Failed to build CodeEntity: {}", e);
-                // Return a minimal valid entity on error
-                // Only set required fields
-                CodeEntityBuilder::default()
-                    .entity_id("error".to_string())
-                    .name("error".to_string())
-                    .qualified_name("error".to_string())
-                    .entity_type(entity_type)
-                    .location(location)
-                    .language(Language::Rust)
-                    .file_path(file_path.to_path_buf())
-                    .line_range((0, 0))
-                    .build()
-                    .unwrap_or_else(|build_err| {
-                        tracing::error!("Failed to build minimal CodeEntity: {}", build_err);
-                        panic!("Cannot create minimal CodeEntity: {}", build_err);
-                    })
-            })
-    }
 }
 
 impl Extractor for RustExtractor {
@@ -107,11 +65,8 @@ impl Extractor for RustExtractor {
         let mut extractor = Self::create_inner_extractor()?;
         let entities = extractor.extract(source, file_path)?;
 
-        // Convert EntityData to CodeEntity
-        Ok(entities
-            .into_iter()
-            .map(|entity| self.convert_to_code_entity(entity, file_path))
-            .collect())
+        // Convert EntityData to CodeEntity using the From trait
+        Ok(entities.into_iter().map(CodeEntity::from).collect())
     }
 }
